Compile to single data.frame

Compiles seperate fst files into one, picking relevant columns

Usage

lfs_compile(
  lfs_directory,
  filter_years = NULL,
  extra_mappings = NULL,
  save_to_folder = FALSE,
  save_variables_report = TRUE,
  aps = FALSE
)

Arguments

lfs_directory

Path to directory with LFS files

filter_years

Vector of years to filter down to

extra_mappings

Either NULL (use default) or a file which has the custom mapping function. See vignette("Adding_Variables", package = "tidylfs")

save_to_folder

If TRUE, will save to the DATA_DIRECTORY environment variable if present

save_variables_report

Save a csv with the list of picked variables?

aps

Annual Population Survey flag, files called eg. "APS 2012.sav"

Value

Nothing - saves the fst file only.