Selects and renames columns from LFS File

Checks which variables are present and picks them in order of priority eg. PWT20 > PWT18 etc. Renames them so that they can be binded to one dataset

Usage

lfs_tidy_file(file, total_files, file_format, extra_mappings = NULL)

Arguments

file

Index of file

total_files

Vector of file names - helper function for cli

file_format

File extension: sav csv or rds

extra_mappings

Either NULL (use default) or an R script with the lfs_extra_mappings function

Value

List of output

Details

To automatically save the compiled file in a directory ready to be loaded with lfs_load(), set the DATA_DIRECTORY environment variable to point at a folder.