Picks column names in order of appearance

When column names change over time, picks the first one that appears from the list in order of appearance. If it doesn't appear, returns a character NA.

Usage

pick_var(vector_of_names, column_names)

Arguments

vector_of_names

character vector of column names

column_names

vector of columns from lfs

Value

First name that appears in cols