Compile to single data.frame

Compiles seperate fst files into one, picking relevant columns

Usage

lfs_compile(
  directory,
  dataset = c("lfs", "aps"),
  years = NULL,
  extra_mappings = NULL,
  aps = lifecycle::deprecated(),
  save_variables_report = lifecycle::deprecated(),
  save_to_folder = lifecycle::deprecated()
)

Arguments

directory

Path to directory with LFS files

dataset

Either "lfs" or "aps".

"lfs" expects files named like "2004 Q1.sav" or "2004 Q2.sav",

"aps" expects files named like "APS 2004.sav" or "APS 2005.sav".

years

Either NULL (all years) or a vector of years to filter by.

extra_mappings

Either NULL (use default) or a function that takes a vector of column names and returns a data frame with lfs_name, new_name, and type. See vignette("Adding_Variables", package = "tidylfs")

aps

Deprecated. Use dataset = "aps" instead.

save_variables_report

Deprecated. Use variable_mapping(lfs) instead.

save_to_folder

Deprecated. Use lfs <- lfs_compile(..), and save manually to a folder.

Value

A data.table with a stacked LFS dataset. Access the variable mapping with variable_mapping(lfs).